Actually? use cache aligned registries. Did not call appropriate function before. Fix wrong pointer arithmetic (compute as byte offset)

+++ include/hip_instrumentation/gpu_queue.hpp
@@ -181,7 +181,7 @@ inline __device__ uint32_t get_cu_id() {
     static constexpr uint32_t SH_PER_SE = 2u;
     static constexpr uint32_t CU_PER_SE = CU_PER_SH * SH_PER_SE;
 
-    uint64_t hw_id = get_hw_id();
+    uint32_t hw_id = get_hw_id();
 
     uint32_t cu_id = (hw_id >> 8) & 0b1111u;
     uint32_t sh_id = (hw_id >> 12) & 0b1;
